+++ src/eml2_3/ActivityTemplate.cpp
@@ -37,7 +37,7 @@ void ActivityTemplate::pushBackParameter(const std::string & title,
 	uint64_t minOccurs, int64_t maxOccurs)
 {
 	// Preconditions
-	if (maxOccurs >= 0 && minOccurs > static_cast<unsigned int>(maxOccurs)) {
+	if (maxOccurs >= 0 && minOccurs > static_cast<uint64_t>(maxOccurs)) {
 		throw invalid_argument("Maximum occurences of a parameter template must be at least equal to the minimum occurences of this same parameter template.");
 	}
